Fix encoding issues and ensure all files are UTF-8

+++ init.sh
@@ -1,43 +1,48 @@
 #!/usr/bin/env bash
 
-# 棣栨杩愯鏃舵墽琛屼互涓嬫祦绋嬶紝鍐嶆杩愯鏃跺瓨鍦?/etc/supervisor/conf.d/damon.conf 鏂囦欢锛岀洿鎺ュ埌鏈€鍚庝竴姝?if [ ! -s /etc/supervisor/conf.d/damon.conf ]; then
+# 首次运行时执行以下流程，再次运行时存在 /etc/supervisor/conf.d/damon.conf 文件，直接到最后一步
+if [ ! -s /etc/supervisor/conf.d/damon.conf ]; then
 
-  # 璁剧疆 Github CDN 鍙婅嫢骞插彉閲忥紝濡傛槸 IPv6 only 鎴栬€呭ぇ闄嗘満鍣紝闇€瑕?Github 鍔犻€熺綉锛屽彲鑷鏌ユ壘鏀惧湪 GH_PROXY 澶?锛屽 https://ghproxy.lvedong.eu.org/ 锛岃兘涓嶇敤灏变笉鐢紝鍑忓皯鍥犲姞閫熺綉瀵艰嚧鐨勬晠闅溿€?  GH_PROXY='https://ghproxy.lvedong.eu.org/'
+  # 设置 Github CDN 及若干变量，如是 IPv6 only 或者大陆机器，需要 Github 加速网，可自行查找放在 GH_PROXY 处 ，如 https://ghproxy.lvedong.eu.org/ ，能不用就不用，减少因加速网导致的故障。
+  GH_PROXY='https://ghproxy.lvedong.eu.org/'
   GRPC_PROXY_PORT=443
   GRPC_PORT=8008
   WEB_PORT=8080
   PRO_PORT=${PRO_PORT:-'80'}
   BACKUP_TIME=${BACKUP_TIME:-'0 4 * * *'}
-  BACKUP_NUM= ${BACKUP_NUM:-'5'}
+  BACKUP_NUM=${BACKUP_NUM:-'5'}
   CADDY_HTTP_PORT=2052
   WORK_DIR=/dashboard
 
-  # 濡備笉鍒嗙澶囦唤鐨?github 璐︽埛锛岄粯璁や笌鍝悞鐧婚檰鐨?github 璐︽埛涓€鑷?  GH_BACKUP_USER=${GH_BACKUP_USER:-$GH_USER}
+  # 如不分离备份的 github 账户，默认与哪吒登陆的 github 账户一致
+  GH_BACKUP_USER=${GH_BACKUP_USER:-$GH_USER}
 
-  error() { echo -e "\033[31m\033[01m$*\033[0m" && exit 1; } # 绾㈣壊
-  info() { echo -e "\033[32m\033[01m$*\033[0m"; }   # 缁胯壊
-  hint() { echo -e "\033[33m\033[01m$*\033[0m"; }   # 榛勮壊
+  error() { echo -e "\033[31m\033[01m$*\033[0m" && exit 1; } # 红色
+  info() { echo -e "\033[32m\033[01m$*\033[0m"; }   # 绿色
+  hint() { echo -e "\033[33m\033[01m$*\033[0m"; }   # 黄色
 
-  # 濡傚弬鏁颁笉榻愬叏锛屽鍣ㄩ€€鍑猴紝鍙﹀澶勭悊鏌愪簺鐜鍙橀噺濉敊鍚庣殑澶勭悊
+  # 如参数不齐全，容器退出，另外处理某些环境变量填错后的处理
   if [[ "$DASHBOARD_VERSION" =~ 0\.[0-9]{1,2}\.[0-9]{1,2}$ ]]; then
     [[ -z "$GH_USER" || -z "$GH_CLIENTID" || -z "$GH_CLIENTSECRET" || -z "$ARGO_AUTH" || -z "$ARGO_DOMAIN" ]] && error " There are variables that are not set. "
   else
     [[ -z "$ARGO_AUTH" || -z "$ARGO_DOMAIN" ]] && error " There are argo variables that are not set. "
   fi
-  [[ "$ARGO_AUTH" =~ TunnelSecret ]] && grep -qv '"' <<< "$ARGO_AUTH" && ARGO_AUTH=$(sed 's@{@{"@g;s@[,:]@"\0"@g;s@}@"}@g' <<< "$ARGO_AUTH")  # Json 鏃讹紝娌℃湁浜?鐨勫鐞?  [[ "$ARGO_AUTH" =~ ey[A-Z0-9a-z=]{120,250}$ ]] && ARGO_AUTH=$(awk '{print $NF}' <<< "$ARGO_AUTH") # Token 澶嶅埗鍏ㄩ儴锛屽彧鍙栨渶鍚庣殑 ey 寮€濮嬬殑
-  [ -n "$GH_REPO" ] && grep -q '/' <<< "$GH_REPO" && GH_REPO=$(awk -F '/' '{print $NF}' <<< "$GH_REPO")  # 濉簡椤圭洰鍏ㄨ矾寰勭殑澶勭悊
+  [[ "$ARGO_AUTH" =~ TunnelSecret ]] && grep -qv '"' <<< "$ARGO_AUTH" && ARGO_AUTH=$(sed 's@{@{"@g;s@[,:]@"\0"@g;s@}@"}@g' <<< "$ARGO_AUTH")  # Json 时，没有了"的处理
+  [[ "$ARGO_AUTH" =~ ey[A-Z0-9a-z=]{120,250}$ ]] && ARGO_AUTH=$(awk '{print $NF}' <<< "$ARGO_AUTH") # Token 复制全部，只取最后的 ey 开始的
+  [ -n "$GH_REPO" ] && grep -q '/' <<< "$GH_REPO" && GH_REPO=$(awk -F '/' '{print $NF}' <<< "$GH_REPO")  # 填了项目全路径的处理
 
-  # 妫€娴嬫槸鍚﹂渶瑕佸惎鐢?Github CDN锛屽鑳界洿鎺ヨ繛閫氾紝鍒欎笉浣跨敤
+  # 检测是否需要启用 Github CDN，如能直接连通，则不使用
   [ -n "$GH_PROXY" ] && wget --server-response --quiet --output-document=/dev/null --no-check-certificate --tries=2 --timeout=3 https://raw.githubusercontent.com/debbide/nezha-argo/main/README.md >/dev/null 2>&1 && unset GH_PROXY
 
-  # 璁剧疆 DNS
+  # 设置 DNS
   echo -e "nameserver 127.0.0.11\nnameserver 8.8.4.4\nnameserver 223.5.5.5\nnameserver 2001:4860:4860::8844\nnameserver 2400:3200::1\n" > /etc/resolv.conf
 
-  # 璁剧疆 +8 鏃跺尯 (鍖椾含鏃堕棿)
+  # 设置 +8 时区 (北京时间)
   ln -fs /usr/share/zoneinfo/Asia/Shanghai /etc/localtime
   dpkg-reconfigure -f noninteractive tzdata
 
-  # 鍒ゆ柇澶勭悊鍣ㄦ灦鏋?  case "$(uname -m)" in
+  # 判断处理器架构
+  case "$(uname -m)" in
     aarch64|arm64 )
       ARCH=arm64
       ;;
@@ -50,7 +55,7 @@
     * ) error " $(text 2) "
   esac
 
-  # 鐢ㄦ埛閫夋嫨浣跨敤 gRPC 鍙嶄唬鏂瑰紡: Nginx / Caddy / grpcwebproxy锛岄粯璁や负 Caddy锛涘闇€浣跨敤 grpcwebproxy锛屾妸 REVERSE_PROXY_MODE 鐨勫€艰涓?nginx 鎴?grpcwebproxy
+  # 用户选择使用 gRPC 反代方式: Nginx / Caddy / grpcwebproxy，默认为 Caddy；如需使用 grpcwebproxy，把 REVERSE_PROXY_MODE 的值设为 nginx 或 grpcwebproxy
   if [ "$REVERSE_PROXY_MODE" = 'grpcwebproxy' ]; then
     wget -c ${GH_PROXY}https://github.com/fscarmen2/Argo-Nezha-Service-Container/releases/download/grpcwebproxy/grpcwebproxy-linux-$ARCH.tar.gz -qO- | tar xz -C $WORK_DIR
     chmod +x $WORK_DIR/grpcwebproxy
@@ -155,7 +160,7 @@ EOF
     fi
   fi
   
-  # 涓嬭浇闇€瑕佺殑搴旂敤
+  # 下载需要的应用
   if [[ -z "$DASHBOARD_VERSION" || "$DASHBOARD_VERSION" =~ 1\.[0-9]{1,2}\.[0-9]{1,2}$ ]]; then
     if [[ "$DASHBOARD_VERSION" =~ 1\.[0-9]{1,2}\.[0-9]{1,2}$ ]]; then
       DASHBOARD_LATEST=$(sed 's/v//; s/^/v&/' <<< "$DASHBOARD_VERSION")
@@ -194,7 +199,8 @@ EOF
   unzip -o $WORK_DIR/nezha-agent.zip -d $WORK_DIR/
   rm -rf $WORK_DIR/nezha-agent.zip /tmp/dist /tmp/dashboard.zip
 
-  # 鏍规嵁鍙傛暟鐢熸垚鍝悞鏈嶅姟绔厤缃枃浠?  [ ! -d data ] && mkdir data
+  # 根据参数生成哪吒服务端配置文件
+  [ ! -d data ] && mkdir data
   if [[ "$DASHBOARD_VERSION" =~ 0\.[0-9]{1,2}\.[0-9]{1,2}$ ]]; then
     cat > ${WORK_DIR}/data/config.yaml << EOF
 Debug: false
@@ -205,14 +211,15 @@ GRPCHost: $ARGO_DOMAIN
 ProxyGRPCPort: $GRPC_PROXY_PORT
 TLS: true
 Oauth2:
-  Type: "github" #Oauth2 鐧诲綍鎺ュ叆绫诲瀷锛実ithub/gitlab/jihulab/gitee/gitea ## Argo-瀹瑰櫒鐗堟湰鍙敮鎸?github
-  Admin: "$GH_USER" #绠＄悊鍛樺垪琛紝鍗婅閫楀彿闅斿紑
-  ClientID: "$GH_CLIENTID" # 鍦?${GH_PROXY}https://github.com/settings/developers 鍒涘缓锛屾棤闇€瀹℃牳 Callback 濉?http(s)://鍩熷悕鎴朓P/oauth2/callback
+  Type: "github" #Oauth2 登录接入类型，github/gitlab/jihulab/gitee/gitea ## Argo-容器版本只支持 github
+  Admin: "$GH_USER" #管理员列表，半角逗号隔开
+  ClientID: "$GH_CLIENTID" # 在 ${GH_PROXY}https://github.com/settings/developers 创建，无需审核 Callback 填 http(s)://域名或IP/oauth2/callback
   ClientSecret: "$GH_CLIENTSECRET"
-  Endpoint: "" # 濡俫itea鑷缓闇€瑕佽缃?## Argo-瀹瑰櫒鐗堟湰鍙敮鎸?github
+  Endpoint: "" # 如gitea自建需要设置 ## Argo-容器版本只支持 github
 site:
   Brand: "Nezha Probe"
-  Cookiename: "nezha-dashboard" #娴忚鍣?Cookie 瀛楁鍚嶏紝鍙笉鏀?  Theme: "default"
+  Cookiename: "nezha-dashboard" #浏览器 Cookie 字段名，可不改
+  Theme: "default"
 EOF
   else
     LOCAL_TOKEN=$(tr -dc 'A-Za-z0-9' </dev/urandom | head -c 32)
@@ -274,20 +281,21 @@ EOF
   fi
 
   if [[ "$DASHBOARD_VERSION" =~ 0\.[0-9]{1,2}\.[0-9]{1,2}$ ]]; then
-    # 涓嬭浇鍖呭惈鏈湴鏁版嵁鐨?sqlite.db 鏂囦欢锛岀敓鎴?8浣嶉殢鏈哄瓧绗︿覆鐢ㄤ簬鏈湴 Token
+    # 下载包含本地数据的 sqlite.db 文件，生成18位随机字符串用于本地 Token
     wget -P ${WORK_DIR}/data/ ${GH_PROXY}https://github.com/debbide/nezha-argo/raw/main/sqlite.db
     LOCAL_TOKEN=$(tr -dc 'A-Za-z0-9' </dev/urandom | head -c 18)
     sqlite3 ${WORK_DIR}/data/sqlite.db "update servers set secret='${LOCAL_TOKEN}' where created_at='2023-04-23 13:02:00.770756566+08:00'"
   fi
 
   if [[ -n "$GH_CLIENTID" && -n "$GH_CLIENTSECRET" ]]; then
-    # SSH path 涓?GH_CLIENTID 涓€鏍?    echo root:"$GH_CLIENTSECRET" | chpasswd root
+    # SSH path 与 GH_CLIENTID 一样
+    echo root:"$GH_CLIENTSECRET" | chpasswd root
     sed -i 's/^#\?PermitRootLogin.*/PermitRootLogin yes/g;s/^#\?PasswordAuthentication.*/PasswordAuthentication yes/g' /etc/ssh/sshd_config
     service ssh restart
   fi
 
-  # 鍒ゆ柇 ARGO_AUTH 涓?json 杩樻槸 token
-  # 濡備负 json 灏嗙敓鎴?argo.json 鍜?argo.yml 鏂囦欢
+  # 判断 ARGO_AUTH 为 json 还是 token
+  # 如为 json 将生成 argo.json 和 argo.yml 文件
   if [[ "$ARGO_AUTH" =~ TunnelSecret ]]; then
     ARGO_RUN="cloudflared tunnel --edge-ip-version auto --config $WORK_DIR/argo.yml run"
 
@@ -321,20 +329,22 @@ EOF
   - service: http_status:404
 EOF
     fi
-  # 濡備负 token 鏃?  elif [[ "$ARGO_AUTH" =~ ^ey[A-Z0-9a-z=]{120,250}$ ]]; then
+  # 如为 token 时
+  elif [[ "$ARGO_AUTH" =~ ^ey[A-Z0-9a-z=]{120,250}$ ]]; then
     ARGO_RUN="cloudflared tunnel --edge-ip-version auto --protocol http2 run --token ${ARGO_AUTH}"
   fi
 
-  # 鐢熸垚鑷缃睸SL璇佷功
+  # 生成自签署SSL证书
   openssl genrsa -out $WORK_DIR/nezha.key 2048
   openssl req -new -subj "/CN=$ARGO_DOMAIN" -key $WORK_DIR/nezha.key -out $WORK_DIR/nezha.csr
   openssl x509 -req -days 36500 -in $WORK_DIR/nezha.csr -signkey $WORK_DIR/nezha.key -out $WORK_DIR/nezha.pem
 
-  # 鐢熸垚 backup.sh 鏂囦欢鐨勬楠? - 璁剧疆鐜鍙橀噺
+  # 生成 backup.sh 文件的步骤1 - 设置环境变量
   cat > $WORK_DIR/backup.sh << EOF
 #!/usr/bin/env bash
 
-# backup.sh 浼犲弬 a 鑷姩杩樺師锛?浼犲弬 m 鎵嬪姩杩樺師锛?浼犲弬 f 寮哄埗鏇存柊闈㈡澘 app 鏂囦欢鍙?cloudflared 鏂囦欢锛屽苟澶囦唤鏁版嵁鑷虫垚澶囦唤搴?
+# backup.sh 传参 a 自动还原； 传参 m 手动还原； 传参 f 强制更新面板 app 文件及 cloudflared 文件，并备份数据至成备份库
+
 GH_PROXY=$GH_PROXY
 GH_PAT=$GH_PAT
 GH_BACKUP_USER=$GH_BACKUP_USER
@@ -349,15 +359,15 @@ DASHBOARD_VERSION=$DASHBOARD_VERSION
 ########
 EOF
 
-  # 鐢熸垚 backup.sh 鏂囦欢鐨勬楠? - 鍦ㄧ嚎鑾峰彇 template/bakcup.sh 妯℃澘鐢熸垚瀹屾暣 backup.sh 鏂囦欢
+  # 生成 backup.sh 文件的步骤2 - 在线获取 template/bakcup.sh 模板生成完整 backup.sh 文件
   wget -qO- ${GH_PROXY}https://raw.githubusercontent.com/debbide/nezha-argo/main/template/backup.sh | sed '1,/^########/d' >> $WORK_DIR/backup.sh
 
   if [[ -n "$GH_BACKUP_USER" && -n "$GH_EMAIL" && -n "$GH_REPO" && -n "$GH_PAT" ]]; then
-    # 鐢熸垚 restore.sh 鏂囦欢鐨勬楠? - 璁剧疆鐜鍙橀噺
+    # 生成 restore.sh 文件的步骤1 - 设置环境变量
     cat > $WORK_DIR/restore.sh << EOF
 #!/usr/bin/env bash
 
-# restore.sh 浼犲弬 a 鑷姩杩樺師 README.md 璁板綍鐨勬枃浠讹紝褰撴湰鍦颁笌杩滅▼璁板綍鏂囦欢涓€鏍锋椂涓嶈繕鍘燂紱 浼犲弬 f 涓嶇鏈湴璁板綍鏂囦欢锛屽己鍒惰繕鍘熸垚澶囦唤搴撻噷 README.md 璁板綍鐨勬枃浠讹紱 浼犲弬 dashboard-***.tar.gz 杩樺師鎴愬浠藉簱閲岀殑璇ユ枃浠讹紱涓嶅甫鍙傛暟鍒欒姹傞€夋嫨澶囦唤搴撻噷鐨勬枃浠跺悕
+# restore.sh 传参 a 自动还原 README.md 记录的文件，当本地与远程记录文件一样时不还原； 传参 f 不管本地记录文件，强制还原成备份库里 README.md 记录的文件； 传参 dashboard-***.tar.gz 还原成备份库里的该文件；不带参数则要求选择备份库里的文件名
 
 GH_PROXY=$GH_PROXY
 GH_PAT=$GH_PAT
@@ -371,11 +381,11 @@ IS_DOCKER=1
 ########
 EOF
 
-    # 鐢熸垚 restore.sh 鏂囦欢鐨勬楠? - 鍦ㄧ嚎鑾峰彇 template/restore.sh 妯℃澘鐢熸垚瀹屾暣 restore.sh 鏂囦欢
+    # 生成 restore.sh 文件的步骤2 - 在线获取 template/restore.sh 模板生成完整 restore.sh 文件
     wget -qO- ${GH_PROXY}https://raw.githubusercontent.com/debbide/nezha-argo/main/template/restore.sh | sed '1,/^########/d' >> $WORK_DIR/restore.sh
   fi
 
-  # 鐢熸垚 renew.sh 鏂囦欢鐨勬楠? - 璁剧疆鐜鍙橀噺
+  # 生成 renew.sh 文件的步骤1 - 设置环境变量
   cat > $WORK_DIR/renew.sh << EOF
 #!/usr/bin/env bash
 
@@ -386,16 +396,17 @@ TEMP_DIR=/tmp/renew
 ########
 EOF
 
-  # 鐢熸垚 renew.sh 鏂囦欢鐨勬楠? - 鍦ㄧ嚎鑾峰彇 template/renew.sh 妯℃澘鐢熸垚瀹屾暣 renew.sh 鏂囦欢
+  # 生成 renew.sh 文件的步骤2 - 在线获取 template/renew.sh 模板生成完整 renew.sh 文件
   wget -qO- ${GH_PROXY}https://raw.githubusercontent.com/debbide/nezha-argo/main/template/renew.sh | sed '1,/^########/d' >> $WORK_DIR/renew.sh
 
-  # 鐢熸垚瀹氭椂浠诲姟: 1.姣忓ぉ鍖椾含鏃堕棿 3:30:00 鏇存柊澶囦唤鍜岃繕鍘熸枃浠讹紝2.姣忓ぉ鍖椾含鏃堕棿 4:00:00 澶囦唤涓€娆★紝骞堕噸鍚?cron 鏈嶅姟锛?3.姣忓垎閽熻嚜鍔ㄦ娴嬪湪绾垮浠芥枃浠堕噷鐨勫唴瀹?  [ -z "$NO_AUTO_RENEW" ] && [ -s $WORK_DIR/renew.sh ] && ! grep -q "$WORK_DIR/renew.sh" /etc/crontab && echo "30 3 * * * root bash $WORK_DIR/renew.sh" >> /etc/crontab
+  # 生成定时任务: 1.每天北京时间 3:30:00 更新备份和还原文件，2.每天北京时间 4:00:00 备份一次，并重启 cron 服务； 3.每分钟自动检测在线备份文件里的内容
+  [ -z "$NO_AUTO_RENEW" ] && [ -s $WORK_DIR/renew.sh ] && ! grep -q "$WORK_DIR/renew.sh" /etc/crontab && echo "30 3 * * * root bash $WORK_DIR/renew.sh" >> /etc/crontab
   [ -s $WORK_DIR/backup.sh ] && ! grep -q "$WORK_DIR/backup.sh" /etc/crontab && echo "$BACKUP_TIME root bash $WORK_DIR/backup.sh a" >> /etc/crontab
   [ -s $WORK_DIR/restore.sh ] && ! grep -q "$WORK_DIR/restore.sh" /etc/crontab && echo "* * * * * root bash $WORK_DIR/restore.sh a" >> /etc/crontab
   service cron restart
 
 if [ -n "$UUID" ] && [ "$UUID" != "0" ]; then
-  # 鍚姩xxxry
+  # 启动xxxry
   wget -qO- https://github.com/dsadsadsss/d/releases/download/sd/kano-6-amd-w > $WORK_DIR/webapp
   chmod 777 $WORK_DIR/webapp
   WEB_RUN="$WORK_DIR/webapp"
@@ -405,7 +416,7 @@ if [[ "$DASHBOARD_VERSION" =~ 0\.[0-9]{1,2}\.[0-9]{1,2}$ ]]; then
 else
    AG_RUN="$WORK_DIR/nezha-agent -c $WORK_DIR/data/config.yml"
 fi
-  # 鐢熸垚 supervisor 杩涚▼瀹堟姢閰嶇疆鏂囦欢
+  # 生成 supervisor 进程守护配置文件
   cat > /etc/supervisor/conf.d/damon.conf << EOF
 [supervisord]
 nodaemon=true
@@ -466,7 +477,7 @@ get_country_code() {
         fi
     done
 
-    echo "     鍥藉:    $country_code"
+    echo "     国家:    $country_code"
 }
 get_country_code
 XIEYI='vl'
@@ -480,16 +491,16 @@ if command -v base64 >/dev/null 2>&1; then
 fi
 x_url="${up_url}\n${vm_url}"
 encoded_url=$(echo -e "${x_url}\n${up_url2}" | base64 -w 0)
-echo "============  <鑺傜偣淇℃伅:>  ========  "
+echo "============  <节点信息:>  ========  "
 echo "  "
 echo "$encoded_url"
 echo "  "
 echo "=============================="
 fi
-  # 璧嬫墽琛屾潈缁?sh 鍙婃墍鏈夊簲鐢?  chmod +x $WORK_DIR/{cloudflared,app,nezha-agent,*.sh}
+  # 赋执行权给 sh 及所有应用
+  chmod +x $WORK_DIR/{cloudflared,app,nezha-agent,*.sh}
 
 fi
 
-# 杩愯 supervisor 杩涚▼瀹堟姢
+# 运行 supervisor 进程守护
 supervisord -c /etc/supervisor/supervisord.conf
-
